amq2api/requirements.txt
# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP 客户端
httpx==0.26.0

# 环境变量管理
python-dotenv==1.0.0

# 数据验证
pydantic==2.5.3

# 日志
python-json-logger==2.0.7

# Token 计数
tiktoken==0.5.2
